deployment Functions

+++ Automated Deployment/AutomatedDeployment.Api/Services/BackupServices.cs
@@ -9,7 +9,7 @@ namespace AutomatedDeployment.Api.Services
 {
     public class BackupServices:IBackupServices
     {
-        
+       
         public void MoveTOBackUpFolder(List<string> filesName, string assemblyPath,string backupPath)
         {
             DirectoryInfo di = new DirectoryInfo(backupPath);
@@ -24,6 +24,7 @@ namespace AutomatedDeployment.Api.Services
             }
             foreach (var item in filesName)
             {
+                if(File.Exists(assemblyPath + '\\' + item))
                 File.Move(assemblyPath+'\\'+item, backupPath + '\\' + item);
             }
